Tidy up thumbnail generation in contributions
