+++ routes/contributions.js
@@ -17,6 +17,7 @@ var db = require('seraph')({
 var _ = require('underscore');
 var mmm = require('mmmagic'),
       Magic = mmm.Magic;
+var contributionUtil = require('./contributionutil');
 
 
 // route: /api/contributions
@@ -295,12 +296,7 @@ router.route('/')
 	 *           If tags are created, contribution creator will be set to this tags as the creator of these tags.
 	 *
 	 */
-	.post(auth.ensureAuthenticated, function(req, res, next) {
-
-		req.tempFileDest = './uploads/users/' + req.user.id + '/temp/' + Date.now();
-		next();
-
-	 }, multer({storage: storage.attachmentStorage}).array('attachments'), function(req, res, next){
+	.post(auth.ensureAuthenticated, contributionUtil.initTempFileDest, multer({storage: storage.attachmentStorage}).array('attachments'), function(req, res, next){
 
 		// Creating the contribution node, then link it to the creator (user)
 		var query = [
@@ -364,64 +360,7 @@ router.route('/')
 			}
 		}); 
 
-	}, function(req, res){
-		
-		if (req.files.length === 0) {
-			res.status(200);
-			return res.send('success');
-		}
-
-		// move the files
-		var tempFileDest = req.tempFileDest;
-		var attachmentsDest = './uploads/contributions/' + req.contributionId + '/attachments/';
-
-		// first create the dir
-		// move the files
-		fs.copy(tempFileDest, attachmentsDest, function(err){
-			if (err) {
-				console.error(err);
-			} else {
-				fs.remove(tempFileDest, function(err){
-					console.log(err);
-				})
-			}
-		});
-
-		var createQueries = req.files.map((f, idx) =>
-			' CREATE (a' + idx + ':attachment {dateUploaded: ' + Date.now() + ', size: ' + f.size + ', name: "' + f.filename + '"})' +
-			' WITH u, c, a' + idx + 
-			' CREATE (u)-[:UPLOADED]->(a' + idx + ')' +
-			' WITH a' + idx + ',c,u' +
-			' CREATE (a' + idx + ')<-[:ATTACHMENT]-(c)'
-		);
-
-		var query = [
-			'MATCH (u:user) WHERE ID(u)={userIdParam}',
-			'WITH u',
-			'MATCH (c:contribution) WHERE ID(c)={contributionIdParam}',
-			'WITH u, c'
-		];
-
-		query = query.concat(createQueries);
-		query = query.join('\n');
-
-		var params = {
-			userIdParam: req.user.id,
-			contributionIdParam: req.contributionId
-		};
-
-		db.query(query, params, function(error, result){
-			if (error){
-				console.log(error);
-				res.status(500);
-				res.send('error in uploading file as attachment');
-			} else {
-				res.status(200);
-				res.send('success');
-			}
-		});
-
-	});
+	}, contributionUtil.updateDatabaseWithAttachmentsAndGenerateThumbnails);
 
 // route: /api/contributions/:contributionId
 router.route('/:contributionId')
@@ -720,165 +659,9 @@ router.route('/:contributionId/rate')
 
 //route: /api/contributions/:contributionId/attachments
 router.route('/:contributionId/attachments')
-	.post(auth.ensureAuthenticated, function(req, res, next){
-		// ensure user owns the contribution id first
-		var query = [
-			'MATCH (c:contribution) WHERE ID(c)={contributionIdParam}',
-			'RETURN c.createdBy as id'
-		].join('\n');
-
-		var params = {
-			contributionIdParam: parseInt(req.params.contributionId)
-		};
-
-		db.query(query, params, function(error, result){
-			if (error){
-				console.log(error);
-				res.send('error');
-			} else {
-				var contributionCreatorId = result[0].id;
-				var userId = parseInt(req.user.id);
-
-				if (userId === contributionCreatorId) {
-					next();
-				} else {
-					res.send('not the owner');
-				}
-			}
-		});
-	}, function(req, res, next) {
-
-		req.tempFileDest = './uploads/users/' + req.user.id + '/temp/' + Date.now();
-		next();
-
-	}, multer({storage: storage.attachmentStorage}).array('attachments'), function(req, res){
-		// add attachments (can have multiple..)
-
-			if (req.files.length === 0) {
-			res.status(200);
-			return res.send('success');
-		}
-
-		// move the files
-		var tempFileDest = req.tempFileDest;
-		var attachmentsDest = './uploads/contributions/' + req.params.contributionId + '/attachments/';
-
-		var transferPromise = new Promise(function(resolve, reject){
-			// can factor this out also..
-			fs.copy(tempFileDest, attachmentsDest, function(err){
-				if (err) {
-					console.error(err);
-				} else {
-					fs.remove(tempFileDest, function(err){
-						resolve();
-					})
-				}
-			});
-		})
-		.then(function(){
-
-			return new Promise(function(resolve, reject){
-				var createQueries = req.files.map((f, idx) =>
-					' CREATE (a' + idx + ':attachment {dateUploaded: ' + Date.now() + ', size: ' + f.size + ', name: "' + f.filename + '", thumb:false })' +
-					' WITH u, c, a' + idx + 
-					' CREATE (u)-[:UPLOADED]->(a' + idx + ')' +
-					' WITH a' + idx + ',c,u' +
-					' CREATE (a' + idx + ')<-[:ATTACHMENT]-(c)'
-				);
-
-				console.log('createQueries: ' + createQueries);
-
-				var query = [
-					'MATCH (u:user) WHERE ID(u)={userIdParam}',
-					'WITH u',
-					'MATCH (c:contribution) WHERE ID(c)={contributionIdParam}',
-					'WITH u, c'
-				];
-
-				var returnQuery = req.files.reduce(function(acc, f, idx){
-					return acc + (idx === 0 ? ' ' : '+') + 'collect(id(a' + idx + '))';
-				}, 'RETURN ');
-
-				returnQuery += ' as id';
-
-				query = query.concat(createQueries, returnQuery);
-				query = query.join('\n');
-
-				console.log(query);
-
-				var params = {
-					userIdParam: req.user.id,
-					contributionIdParam: parseInt(req.params.contributionId)
-				};
-
-				db.query(query, params, function(error, result){
-					if (error){
-						console.log(error);
-						res.status(500);
-						res.send('error in uploading file as attachment');
-					} else {
-						res.status(200);
-						res.send('success');
-						resolve(result[0]);
-					}
-				});
-			});
-
-		})
-		.then(function(idArray){
-			// handle the thumbnail generation here
-			var magic = new Magic(mmm.MAGIC_MIME_TYPE);
-		  req.files.map((f, idx) => {
-		  	magic.detectFile(attachmentsDest + f.filename, function(err, result) {
-		      if (err) {
-		      	console.log(err);
-		      	return;
-		      }
-
-		      var isImage = result.split("/")[0] === "image";
-		      console.log(isImage);
-		      if (!isImage) {
-		      	return;
-		      }
-
-		      // create the /thumbnails folder if not exist yet
-		      mkdirp(attachmentsDest + '/thumbnails/', function(err){
-		      	if (err)
-		      		console.log(err);
-		      });
-
-			    gm(attachmentsDest + f.filename)
-				  .resize(300, 300, '^')
-				  .gravity('Center')
-				  .crop(200, 200)
-				  .quality(100)
-				  .write(attachmentsDest + '/thumbnails/' + f.filename, function(error){
-				  	if (error) 
-				  		console.log('error!!! for : ' + f.filename);
-				  	else{
-				  		console.log('found ', idx);
-
-				  		var query = [
-				  			'MATCH (a:attachment) WHERE ID(a)={attachmentIdParam}',
-				  			'SET a.thumb = true'
-				  		].join('\n');
-
-				  		var params = {
-				  			attachmentIdParam: idArray[idx],
-				  		}
-
-				  		db.query(query, params, function(error, result){
-				  			if (error)
-				  				console.log(error);
-				  		});
-				  	}
-				  });
-				});
-		  })
-
-		});
-
-	})
+	.post(auth.ensureAuthenticated, contributionUtil.ensureUserOwnsContribution, contributionUtil.initTempFileDest, 
+	multer({storage: storage.attachmentStorage}).array('attachments'), 
+	contributionUtil.updateDatabaseWithAttachmentsAndGenerateThumbnails);
 
 //route: /api/contributions/:contributionId/attachments/:attachmentId
 router.route('/:contributionId/attachments/:attachmentId')
