+++ views/upload.ejs
@@ -4,10 +4,10 @@
 	<title>test upload</title>
 </head>
 <body>
-	<form enctype="multipart/form-data" action="/api/contributions/485/attachments" method="post">
+	<form enctype="multipart/form-data" action="/api/contributions/528/attachments" method="post">
 		<input type="text" name="tags" id="tags" value="funny" />
-		<input type="text" name="title" id="title" value="test upload" />
-		<input type="text" name="body" id="body" value="test upload" />
+		<input type="text" name="title" id="title" value="thumbnail test" />
+		<input type="text" name="body" id="body" value="thumbnail test" />
 		<input type="text" name="ref" id="ref" value="374" />
 		<input type="text" name="contentType" value="test" />
 	  <input type="file" name="attachments" multiple />
